Purchase Order id added. Receipt Status field of Response is mapped to business  status code field of Cosmos. ReceiptNumbers data type changed from Integer to Long.

+++ src/main/java/com/walmart/finance/ap/fds/receiving/response/ReceivingInfoResponse.java
@@ -20,7 +20,7 @@ public class ReceivingInfoResponse {
 
     private String purchaseOrderId;
 
-    private Integer receiptNumber;
+    private Long receiptNumber;
 
     private Integer transactionType;
